Move container into layout and get rid of the concept of sections
